+++ core_js/storage.js
@@ -132,6 +132,12 @@ function initStorage(items)
 
     // Start the clearurls.js
     start();
+
+    // Start the context_menu
+    contextMenuStart();
+
+    // Start history listener
+    historyListenerStart();
 }
 
 /**
@@ -153,6 +159,8 @@ function initSettings()
     storage.hashURL = "https://gitlab.com/KevinRoebert/ClearUrls/-/jobs/artifacts/master/raw/rules.min.hash?job=hash%20rules";
     storage.ruleURL = "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/data.min.json";
     storage.reportServer = "https://clearurls.xn--rb-fka.it";
+    storage.contextMenuEnabled = true;
+    storage.historyListenerEnabled = true;
 
     if(getBrowser() === "Firefox") {
         storage.types = ["font", "image", "imageset", "main_frame", "media", "object", "object_subrequest", "other", "script", "stylesheet", "sub_frame", "websocket", "xbl", "xml_dtd", "xmlhttprequest", "xslt"];
@@ -169,9 +177,9 @@ function replaceOldURLs(url)
 {
     switch (url) {
         case "https://raw.githubusercontent.com/KevinRoebert/ClearUrls/master/data/rules.hash?flush_cache=true":
-        return "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/rules.hash";
+        return "https://gitlab.com/KevinRoebert/ClearUrls/-/jobs/artifacts/master/raw/rules.min.hash?job=hash%20rules";
         case "https://raw.githubusercontent.com/KevinRoebert/ClearUrls/master/data/data.json?flush_cache=true":
-        return "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/data.json";
+        return "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/data.min.json";
         case "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/rules.hash":
         return "https://gitlab.com/KevinRoebert/ClearUrls/-/jobs/artifacts/master/raw/rules.min.hash?job=hash%20rules";
         case "https://gitlab.com/KevinRoebert/ClearUrls/raw/master/data/data.json":